requirements.txt
fastapi==0.110.1
uvicorn[standard]
python-multipart
numpy

# nsfw
tensorflow>=2.2.0;sys_platform != 'darwin'
tensorflow_macos>=2.5.0;sys_platform == 'darwin'
tensorflow-hub==0.12.0
pillow